fix: Backport ICU support for placeholder Japanese New Era name
This change ports changes from the ICU 62.1 maint branch to the Chromium 69 ICU, in order to enable placeholder name support for the New Japanese Era.

Note that this change required re-generating the checked-in Chromium ICU data files (*.dat) which is why the .patch file is so large.

Also included in this patch are some minor changes to the Chromium scripts so that the build process actually works on Ubuntu.